api/requirements.txt
Django==1.11.6
psycopg2==2.7.3.1
pytz==2017.2
djangorestframework==3.7.0
django-rest-marshmallow==3.0.0
marshmallow==2.13.6
